feat(arena): 3x arena health + UESP-calibrated damage model
- Apply CHEAT_BASE_HEALTH_MULTIPLIER=3 to max Health at Fighter creation (arena);
  Stamina/Magicka un-multiplied. Fighter holds raw HP (u32) derived from character
  level (health_for_level); damage applies to raw HP.
- Wire ReceiveDamage stats pack a FRACTION of max (round(cur/max x1023)) in the
  HIGH 32 bits, sequenceId in the LOW 32 — fixes the PackedStats half-swap (the
  first-pass RE had them backwards; a full actor reads 1023 from the high half,
  verified vs 4330 captured frames).
- UESP-derived DamageModel (combat/tables.rs): additive weapon surface
  (15*(smithy+1)+quality_bonus x weight), crit/combo swing tiers, spell base-by-rank;
  enchant magnitude calibrated to captured shock/poison tracks; loadout derives a
  level-appropriate Heavy weapon base.

Validated vs all 7 decrypted combat sessions (4330 ReceiveDamage frames):
sum-invariant 100%, enchant additive, Middle=heaviest, wire=fraction. Swing
multipliers not exactly recoverable (post-armor/resist) -> UESP crit/combo used as a
grounded approximation. 41 server tests green.

+++ server/src/arena/combat/resolve.rs
@@ -136,7 +136,7 @@ fn emit_damage(
     resolved: &ResolvedDamage,
 ) -> Vec<(usize, Vec<u8>)> {
     let mut out = Vec::new();
-    combat.fighters[target_slot].take_damage(resolved.total.round().max(0.0) as u16);
+    combat.fighters[target_slot].take_damage(resolved.total.round().max(0.0) as u32);
 
     let msg = {
         let damaged = &combat.fighters[target_slot];
